part5/Notes_BE: add token extractor, user extractor, refactor login route to return user._id

+++ part4/Notes_BE/controllers/notes.js
@@ -1,22 +1,9 @@
 // route logic only (request handlers)
 // handles requests, does not know how the app is configured, just the business logic related to notes
 
-const jwt = require('jsonwebtoken')
 const notesRouter = require('express').Router()
-const logger = require('../utils/logger')
 const Note = require('../models/note')
-const User = require('../models/user')
-
-// creates a helper function
-// extract JWT token from request headers
-const getTokenFrom = request => {
-  const authorization = request.get('authorization') // reads the HTTP header, might contain 'Bearer xxxxxx.yyyyyy.zzzzzz' which means the holder of this token is authenticated
-  if(authorization && authorization.startsWith('Bearer ')) {
-    // transforms authorization without the 'Bearer'
-    return authorization.replace('Bearer ', '')
-  }
-  return null
-}
+const middleware = require('../utils/middleware')
 
 notesRouter.get('/', async (request, response) => {
   const notes = await Note.find({})
@@ -38,24 +25,10 @@ notesRouter.get('/:id', async (request, response) => {
     }
 })
 
-notesRouter.post('/', async (request, response) => {
+notesRouter.post('/', middleware.userExtractor, async (request, response) => {
   const body = request.body
 
-  // jwt.verify returns the original payload (contains the user ID stored in the token)
-  const decodedToken = jwt.verify(getTokenFrom(request), process.env.SECRET)
-
-  if(!decodedToken.id) {
-    return response.status(401).json({ error: 'token invalid' })
-  }
-
-  const user = await User.findById(decodedToken.id)
-  // const user = await User.findById(body.userId)
-
-  console.log('body.userId', body.userId)
-
-  if(!user) {
-    return response.status(400).json({ error: 'UserId missing or not valid' })
-  }
+  const user = request.user
 
   const note = new Note({
     content: body.content,
@@ -67,19 +40,30 @@ notesRouter.post('/', async (request, response) => {
   user.notes = user.notes.concat(savedNote._id)
   await user.save()
   response.status(201).json(savedNote)
-
-  // without async await
-  // note.save()
-  //   .then(savedNote => {
-  //     response.status(201).json(savedNote)
-  //   })
-  //   .catch(error => next(error))
 })
 
-notesRouter.delete('/:id', async (request, response) => {  
+notesRouter.delete('/:id', middleware.userExtractor, async (request, response) => {  
   const id = request.params.id
 
+  const user = request.user
+
+  const note = await Note.findById(id)
+
+  if(!note) {
+    return response.status(400).json({
+      error: 'Note not found!'
+    })
+  }  
+
+  if(note.user.toString() !== user._id.toString()) {
+    return response.status(403).json({ error: 'Unauthorized to delete this note' })
+  }
+
   await Note.findByIdAndDelete(id)
+
+  user.notes = user.notes.filter(note => note.toString() !== id)
+  await user.save()
+
   response.status(204).end()
 
   // without async await
